dossier_ecf/requirements.txt

requests
beautifulsoup4
pandas
sqlalchemy
psycopg2-binary
minio
openpyxl
python-dotenv
lxml